Handle multiple install calls (#125)
It is possible to invoke `office-addin-dev-certs install` more than once at the same time, for example using `npm run dev-server` and `npm run build` which both trigger this, and each may prompt to install the CA certificate. It can be any set of operations but this particular one is the most likely.

This changes the code to check for situations that may arise. For example, after prompting to install the CA certificate, check whether it is already installed before trying to install it. This protects against another instance installing it while the prompt was being shown.