Ensure prompt to install dev certs is visible on non-Windows computers
On non-Windows computers, the detached process that is created to run the dev server does not have a window,
so if the dev server prompts to install dev certs, the prompt is not seen and it will timeout.

To workaround this, prompt if neccessary to install the dev certs before starting the dev server process.