office-addin-debugging: stop command will stop the dev-server if started using the start command (#161)
If "office-addin-debugging start" command starts the dev server, it will save the process id to a json file in the temp folder. When "office-addin-debugging stop" is executed, if the process id was saved, it will stop this process. Thus, "start" and "stop" can start and then stop the dev-server. However, if the dev-server is already running when "start" is called, "stop" will let it keep running.