workaround lerna bootstrap issue
For a clean git repository, running "lerna bootstrap" is failing to link
packages in this repository. Running "npm ls" in "packages\office-addin-manifest" will
have UNMET dependencies for office-addin-cli and office-addin-usage-data.

I can consistently reproduce this, and even going back to older versions of lerna do not resolve it.

The only reliable fix I have found, is to run "npm install" for each of the packages after the "npm install" in the root, and then afterward run "lerna bootstrap".

This updates the "postinstall" command to do that.