Address recent drop in usage data because usagelevel set to off
- The fix is to ensure office-addin-scripts packages don't write out the office-addin-usage-data file before Yo Office does

Yo office imports other packages such as office-addin-manifest, which when loaded as part of an import, invoke the following to create a new usageData object:

export const usageDataObject: usageData.OfficeAddinUsageData = new usageData.OfficeAddinUsageData({
    projectName: "office-addin-manifest",
    instrumentationKey:     usageData.instrumentationKeyForOfficeAddinCLITools,
    raisePrompt: false

});

In the constructor for the office-addin-usage-data, we have the following bit of code:

      if (!this.options.isForTesting && this.options.raisePrompt && jsonData.needToPromptForUsageData(this.options.groupName)) {
        this.usageDataOptIn();
      } else {
        jsonData.writeUsageDataJsonData(this.options.groupName, this.options.usageDataLevel);
      }

Because raisePrompt is set to false, we end up calling jsonData.writeUsageDataJsonData with the usageDataLevel turned to off, and write out the office-addin-usage-data file with the setting. Once this file is written, yo office will see its written and use these settings moving forward.