add option to open dev tools when debugging
This is being added to support the scenario where a developer wants to use the browser dev tools for debugging, but the add-in is running in the background, so there are no controls available to open the dev tools from the webview.

The office-addin-dev-settings "debugging" command has a new option "--open-dev-tools" which will write a registry value ("OpenDevTools" REG_DWORD 1) when present.

This is implemented by adding an optional parameter to the enableDebugging() function, openDevTools: boolean = false, which will write the registry value when true, or remove it (if present) when false.

The office-addin-debugging "start" command has a new option "--dev-tools" which will specify true for the openDevTools parameter when calling devSettings.enableDebugging().

This will only have an effect when supported.

As Office adds support for the Edge Chromium webview (WebView2) on Windows, this will allow the dev tools to be opened (if desired) when using that webview.