custom functions metadata generation doesn't require file output (#397)
The custom functions metadata plugin was always writing to an output file, but it doesn't need to do so. It just needs to emit the json metadata and webpack can either write it out to file or serve it up when using webpack-dev-server.

I have made a breaking change to custom-functions-metadata where the generate() function is renamed to generateCustomFunctionsMetadata() and does not take an output file path.

The custom-functions-metadata CLI command generate still supports an output file path, but it is now optional. If provides, it writes to the file, and if not, the metadata is output to the console.