Removing telemetry from Office-addin-mock (#534)
The telemetry was reporting cycles, given it was reporting for each object created with the test tool.
Jest didn't like this, because apparently, it lead to some memory leaks. So we needed to remove this telemetry to be able to have Jest tests.

For now, I decided that the telemetry is not logging useful data (it is simply logging when someone is using the API, which we can get from the npm website). This data would also be repetitive, so no value in there.