office-addin-dev-certs: add --domains option to install
This allows the set of domains to be customized if needed.

An example where customization is useful is if "localhost" is not allowed as a domain name. You use map a different name to the local port, but the SSL certificate needs to have that name listed.

For example, on Windows, the hosts file can map "dev-host" to 127.0.0.1 (IPv4) and ::1 (IPv6) by adding these lines to %systemroot%\system32\drivers\etc\hosts:

127.0.0.1 	dev-host
::1		dev-host

Then, use the --domains option to specify the IP addresses and host names:

office-addin-dev-certs install --domains 127.0.0.1,::1,localhost,dev-host

Opening the browser to https://dev-host will show that the SSL certificate is valid.